refactor(fantasy): split all services into feature module (queries/mutations/admin)
Move 8 fantasy service files from src/lib/services/ to src/features/fantasy/services/
with read/write separation. Old files become re-export bridges for backward compatibility.

Split files:
- events.ts -> events.queries.ts + events.mutations.ts
- lineups.ts -> lineups.queries.ts + lineups.mutations.ts
- scoring.ts -> scoring.queries.ts + scoring.admin.ts
- predictions.ts -> predictions.queries.ts + predictions.mutations.ts

Verbatim copies:
- fixtures.ts, chips.ts, wildcards.ts
- fantasyLeagues.ts -> leagues.ts
